ESLint: fix errors in components
