+++ src/components/Button/index.js
@@ -1,1 +1,1 @@
-export { default } from './Button';
+export { default } from './Button.jsx';
