+++ src/components/ConfirmDialog/index.js
@@ -1,1 +1,1 @@
-export { default } from './ConfirmDialogContainer';
+export { default } from './ConfirmDialogContainer.jsx';
