+++ src/components/Help/Help.test.jsx
@@ -1,7 +1,4 @@
 import React from 'react';
-import sinon from 'sinon';
-import { expect } from 'chai';
-import Help from './Help.jsx';
 
 describe('Help', () => {
     it('should do something', () => {
