+++ src/components/Icon/index.js
@@ -1,1 +1,1 @@
-export { default } from './Icon';
+export { default } from './Icon.jsx';
