+++ src/components/ItemForm/index.js
@@ -1,1 +1,1 @@
-export { default } from './ItemForm';
+export { default } from './ItemForm.jsx';
