+++ src/components/ItemsList/index.js
@@ -1,1 +1,1 @@
-export { default } from './ItemsList';
+export { default } from './ItemsList.jsx';
