+++ src/components/Link/index.js
@@ -1,1 +1,1 @@
-export { default } from './Link';
+export { default } from './Link.jsx';
