+++ src/components/PasswordInput/index.js
@@ -1,1 +1,1 @@
-export { default } from './PasswordInput';
+export { default } from './PasswordInput.jsx';
