+++ src/components/Redirect/index.js
@@ -1,1 +1,1 @@
-export { default } from './Redirect';
+export { default } from './Redirect.jsx';
