+++ src/components/Search/index.js
@@ -1,1 +1,1 @@
-export { default } from './SearchContainer';
+export { default } from './SearchContainer.jsx';
