+++ src/components/TextArea/TextArea.test.jsx
@@ -1,7 +1,4 @@
 import React from 'react';
-import sinon from 'sinon';
-import { expect } from 'chai';
-import TextArea from './TextArea.jsx';
 
 describe('TextArea', () => {
     it('should do something', () => {
