+++ src/components/TextInput/index.js
@@ -1,1 +1,1 @@
-export { default } from './TextInput';
+export { default } from './TextInput.jsx';
