+++ src/components/Title/index.js
@@ -1,1 +1,1 @@
-export { default } from './Title';
+export { default } from './Title.jsx';
